refactor: move TeamStatsCalculator + StrengthOfScheduleCalculator into BasketballStats
Tier-1 namespace relocation 2.30: consolidates Statistics\TeamStatsCalculator
and StrengthOfSchedule\StrengthOfScheduleCalculator into classes/BasketballStats,
sweeping 6 caller files and regenerating both PHPStan baselines.
