+++ ibl5/phpunit.xml
@@ -145,9 +145,6 @@
         <testsuite name="FranchiseHistory Module Tests">
             <directory>tests/FranchiseHistory</directory>
         </testsuite>
-        <testsuite name="Statistics Module Tests">
-            <directory>tests/Statistics</directory>
-        </testsuite>
         <testsuite name="Scripts Module Tests">
             <directory>tests/Scripts</directory>
         </testsuite>
@@ -207,9 +204,6 @@
         </testsuite>
         <testsuite name="Mail Module Tests">
             <directory>tests/Mail</directory>
-        </testsuite>
-        <testsuite name="StrengthOfSchedule Module Tests">
-            <directory>tests/StrengthOfSchedule</directory>
         </testsuite>
         <testsuite name="JsbParser Module Tests">
             <directory>tests/JsbParser</directory>
