+++ ibl5/tests/UpdateAllTheThings/PowerRankingsUpdaterTest.php
@@ -6,7 +6,7 @@ namespace Tests\UpdateAllTheThings;
 
 use PHPUnit\Framework\TestCase;
 use Updater\PowerRankingsUpdater;
-use Statistics\TeamStatsCalculator;
+use BasketballStats\TeamStatsCalculator;
 use Season\Season;
 use Tests\WideUnit\Mocks\MockDatabase;
